requirements.txt
numpy==1.26.0
Pillow==10.0.1
streamlit==1.27.1
tensorflow==2.14.0
opencv-python